feat: add star rating to extensions for assertion
