+++ server/src/metakb/repository/queries/update_assertion_properties.cypher
@@ -1,3 +1,4 @@
 MATCH (s:Statement {id: $statement_id})
 SET s.direction = $direction
+SET s.extensions = $extensions
 RETURN s
