+++ server/src/metakb/services/manage_data.py
@@ -92,10 +92,15 @@ def _check_for_assertion_updates(
     db_assertion = repository.get_statement(assertion.id)
     if not db_assertion:
         return
-    if db_assertion.strength != assertion.strength:
-        # update assertion in place with new strength value
+    if db_assertion.hasEvidenceLines != assertion.hasEvidenceLines:
+        # update assertion in place with newly-aggregated evidence and derived fields
         merge_assertions(assertion, db_assertion)
         repository.update_assertion_strength(assertion.id, assertion.strength)
+        repository.update_assertion_properties(
+            assertion.id,
+            assertion.direction,
+            assertion.extensions,
+        )
 
 
 def load_from_json(
